# specs/001-ai-inventory-voice/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Inventory Management System API
  description: |
    RESTful API for AI-powered inventory management with voice interaction capabilities.
    Multi-tenant, role-based access control, real-time updates.
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: https://api.example.com/v1
    description: Production server
  - url: https://staging-api.example.com/v1
    description: Staging server
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token with tenant_id and user roles
  schemas:
...
    Recommendation:
      type: object
      required:
        - recommendation_type
        - product_id
        - warehouse_id
        - recommended_value
      properties:
        id:
          type: string
          format: uuid
        recommendation_type:
          type: string
          enum: [reorder_point, reorder_quantity, purchase_order]
        product_id:
          type: string
          format: uuid
        warehouse_id:
          type: string
          format: uuid
        recommended_value:
          type: number
          format: double
        current_value:
          type: number
          format: double
        urgency_score:
          type: number
          format: double
          minimum: 0
          maximum: 100
        confidence_score:
          type: number
          format: double
          minimum: 0
          maximum: 100
        explanation:
          type: string
        status:
          type: string
          enum: [active, approved, rejected, superseded]
        created_at:
          type: string
          format: date-time
...
paths:
...
  /recommendations:
    get:
      tags: [Recommendations]
      summary: Get AI recommendations
      parameters:
        - name: recommendation_type
          in: query
          schema:
            type: string
            enum: [reorder_point, reorder_quantity, purchase_order]
        - name: status
          in: query
          schema:
            type: string
            enum: [active, approved, rejected]
        - name: min_urgency
          in: query
          schema:
            type: integer
            minimum: 0
            maximum: 100
      responses:
        '200':
          description: List of recommendations
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Recommendation'
  /recommendations/{recommendation_id}/explanation:
    get:
      tags: [Recommendations]
      summary: Get detailed explanation for recommendation
      parameters:
        - name: recommendation_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Detailed explanation
          content:
            application/json:
              schema:
                type: object
                properties:
                  explanation:
                    type: string
                  features:
                    type: object
                  confidence_breakdown:
                    type: object
...